requirements.txt
openai>=1.30.0
wikipedia
python-dotenv
